Better python binary detection
Instead of using `AC_PATH_PROG` to find the python binary (which can only find
binaries named `python`, not `python3` and similar), use `AM_PATH_PYTHON`
instead, and take `--with-python` into account as well.

For this to work, we also need to call a few of our python scripts through the
interpreter, to avoid having to generate the scripts.

Fixes #2090.

Signed-off-by: Gergely Nagy <algernon@balabit.com>